requirements.txt
asgiref==3.9.2
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
dj-database-url==3.0.1
Django==5.2.6
django-extensions==4.1
google-auth==2.41.1
gunicorn==23.0.0
idna==3.10
MarkupSafe==3.0.3
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
rsa==4.9.1
sendgrid==6.12.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
